Refactor project CLI into submodule
